refactor(lessons): TeacherLessons sous 300 lignes — éclatement complet (H4)
1091→242 lignes. Extraction:
- composables/useTeacherLessons.js (226): leçons+matières (cache+refresh), filtres, stats, navigation
- composables/useLessonForm.js (115): formulaire modale (save/create/edit/close), SRP, ref lessons injectée
- LessonsFilters.vue, LessonsStatsGrid.vue, TeacherLessonCard.vue, LessonFormModal.vue, LessonsEmptyState.vue (CSS verbatim + @media distribués)
Parité: mêmes appels services, mêmes classes/textes/routes; test G5 existant (shallowMount) inchangé, vert.

Dettes pré-existantes CONSERVÉES à l'identique (relocalisées verbatim, non corrigées):
- modale création/édition sans déclencheur dans la vue (showCreateModal jamais activé)
- fonctions editLesson/deleteLesson commentées (déplacées dans useLessonForm)
- CSS .btn-view minifié sur une ligne + .btn-edit:hover orphelin (dans TeacherLessonCard)
- classes CSS mortes .btn-create / .btn-create-empty conservées

Tests: composable list 4 + composable form 3 + card 3 + filters 2 = 12 verts (+ G5).
